refactor(front): 型・定数を src/types と src/constants へ分離
lib/ に混在していた型（VideoItem / Category / SortOption / Screen）と
定数（CATEGORIES）を、純粋な型は src/types、ランタイム定数は src/constants
へ分離。フォルダ名と中身（型 / 値）を一致させ lib/ はロジック層に専念させる。

- 参照 16 ファイルを @/types / @/constants へ張り替え
- atomic-design-plan.md の構成図に types/ constants/ を明記
- 機能変更なし（typecheck / lint / test 126 / format:check 全通過）
